Fix parsing numbers when they are comma separated Cldr.Number.Parser.scan/2 localises the regex that splits the string into what may be numbers but it wasn't previously capable of handling a comma-separated list of numbers.
The revised regex allows "," (localised grouping symbol)
in a number as long as there is a digit on either side
of it.

Therefore "55, 66" with scan as two number "55" and "66"
whereas "55,66" will scan as "5566" in locales where ","
is a grouping character and "55.55" in locales where it
is a decimal separator.

Thanks for @DaTrader for the bug report. Closes #19.